# Conflicts:
#	team/src/main/java/sansam/team/team/command/application/controller/TeamMemberController.java
#	team/src/main/java/sansam/team/team/command/application/service/TeamMemberService.java
#	team/src/main/java/sansam/team/team/command/application/service/TeamReviewService.java
#	team/src/main/java/sansam/team/team/command/application/service/TeamService.java

+++ team/src/main/java/sansam/team/buildingrule/command/domain/repositroy/BuildingRuleRepository.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 
 public interface BuildingRuleRepository {
     BuildingRule save(BuildingRule buildingRule);
-    Optional<BuildingRule> findById(int ruleSeq);
-    void deleteById(int ruleSeq);
+    Optional<BuildingRule> findById(long ruleSeq);
+    void deleteById(long ruleSeq);
 }
